requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pdfplumber==0.10.3
chromadb==0.4.24 